refactor: split oversized modules to honor the repo's own 400-line budget
- orchestration/workflow.py (834) -> workflow.py (engine, 182) +
  steps_build.py + steps_ship.py + context.py (RunContext, signals,
  escalation machinery) + render.py (artifact markdown)
- stacks/stdlib_code.py (612) -> stdlib_code.py (storage/auth/init) +
  stdlib_api.py (api/server/readme)
- stacks/stdlib_tests.py (585) -> stdlib_tests.py (storage/auth tests,
  shared payload helpers) + stdlib_tests_api.py (API tests)

No behavior change: 151 tests green, mypy --strict clean. Largest
source file is now 392 lines — the budget docs/technical-requirements.md
declares is enforced again true of this repo itself.
